Change to a flag from flaky time.sleeps.
This makes a few assumptions, namely that you're only testing on the
first value returned, anything beyond that will require more powerful
stiff or some personalization.

Also yield from is there to await the coroutine that is being used, if
that isn't there, the sleep while loop will just block the thread
forever and the callback will never get the chance to run. This way it
doesn't.