Remove symlink error with catkin_make.
Because of some strange issue with the install style of python2.7 on
ubuntu, it creates a `local` directory, which contains symlinks to every
file in the parent directory (buzzmobile). One of those child
directories is the `buzzmobile` ros package.

This led to an error when running `catkin_make`, that would say
something like "multiple packages found with the name buzzmobile". This
was resolved by removing the symlink to the child buzzmobile directory
in the install script before calling catkin_make.