Tests should no longer be flaky.
Thi finally makes everything work!

Creating listener and contoller allows us to run the nodes in background
processes, which doesn't run into rospy issues thankfully. There are a
few changes that need to be made w.r.t. thread safety, currently I do a
hack where I set the environmental variable, which needs to be done, but
that is also used later, and I'd much rather explicitly pass the port to
the context managers than use the shell to get it there if we end up
with multiple processes, although IDK if it will matter.